refactor(agentic): split three modules out to stay under the 500-line cap
The docs this change adds pushed `mod.rs` from exactly 500 lines to 612 and
`score.rs` from 496 to 537, and `Enforce ≤500 LoC per source file` is a required
check. Both files were already sitting ON the ceiling, so any addition had to be
paid for by a split. Extracted along the seams the repo already uses — `score.rs`
itself was carved off `mod.rs` for this same reason:

  params.rs   the ParamSpec list        mod.rs   612 -> 431
  render.rs   table() and summary()
  verdict.rs  the PASS/FAIL decision    score.rs 537 -> 465
              plus total_turns / seconds_per_turn

The seam in each case is a real one, not a line-count convenience: `params` is
configuration, `render` decides nothing, and `verdict` judges a tier where the
rest of `score` measures one. Bodies are unchanged; only their home moved, and
`pub mod score` stayed public (a `mod score;` slip while editing the declaration
list was caught by the build).

★ This invalidates the eight records committed in a0f73a5477 — the closure
covers `crates/`, and a pure code move is still a code move. They are re-cut at
this sha rather than left to look valid. The two BFCL legs in flight were killed
mid-run for the same reason; letting them finish would have bought a record
nothing could use.

Workspace clippy, tests and fmt clean.
